Inline ReadOnlyWalDirectoryPolicy as nested class
The 15-line WalDirectoryPolicy stub doesn't warrant its own file. Move
it to a private static nested class inside WalToParquet so the utility
lives in a single source file, matching the spirit of the other
single-file utilities in the cliutil package.
