--- utils/src/main/java/io/questdb/cliutil/ReadOnlyWalDirectoryPolicy.java
@@ -1,50 +1,0 @@
-/*+*****************************************************************************
- *     ___                  _   ____  ____
- *    / _ \ _   _  ___  ___| |_|  _ \| __ )
- *   | | | | | | |/ _ \/ __| __| | | |  _ \
- *   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
- *    \__\_\\__,_|\___||___/\__|____/|____/
- *
- *  Copyright (c) 2014-2019 Appsicle
- *  Copyright (c) 2019-2026 QuestDB
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- ******************************************************************************/
-
-package io.questdb.cliutil;
-
-import io.questdb.cairo.wal.WalDirectoryPolicy;
-import io.questdb.std.str.Path;
-
-// No-op policy for read-only access. TableTransactionLogV2 only invokes
-// these methods on the write path, so all read paths see a no-op stub.
-final class ReadOnlyWalDirectoryPolicy implements WalDirectoryPolicy {
-    @Override
-    public void initDirectory(Path dirPath) {
-    }
-
-    @Override
-    public boolean isInUse(Path path) {
-        return false;
-    }
-
-    @Override
-    public void rollbackDirectory(Path path) {
-    }
-
-    @Override
-    public boolean truncateFilesOnClose() {
-        return false;
-    }
-}
